+++ launchDaemonManagement.zsh
@@ -21,12 +21,13 @@
 #
 # HISTORY
 #
-# Version 2.2.0b7, 16-Dec-2025, Dan K. Snelson (@dan-snelson)
-#   - :warning: Added TelemetryDeck integration to collect anonymous usage data (set `telemetryEnabled="NO"` in `reminderDialog.zsh` to disable) :warning:
+# Version 2.2.0b8, 16-Dec-2025, Dan K. Snelson (@dan-snelson)
+#   - :warning: Added TelemetryDeck integration to collect anonymous usage data (set `telemetryEnabled="NO"` in `reminderDialog.zsh` to disable)
 #   - Addressed Feature Request: Intelligently display reminder dialog after rebooting #42
 #   - Added instructions for monitoring the client-side log
 #   - `assemble.zsh` now outputs to `Artifacts/` (instead of `Resources/`)
 #   - Updated `Resources/sample.plist` to address Feature Request #43
+#   - Harmonized Organization Variables between `launchDaemonManagement.zsh` and `reminderDialog.zsh`
 #
 ####################################################################################################
 
@@ -41,7 +42,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.2.0b7"
+scriptVersion="2.2.0b8"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"
@@ -64,7 +65,7 @@ resetConfiguration="${4:-"All"}"
 # Organization Variables
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-# Organization's Reverse Domain Name Notation (i.e., com.company.division)
+# Organization's Reverse Domain Name Notation (i.e., com.company.division; used for plist domains)
 reverseDomainNameNotation="org.churchofjesuschrist"
 
 # Script Human-readabale Name
@@ -176,6 +177,17 @@ function resetConfiguration() {
         "Uninstall" )
 
             warning "*** UNINSTALLING ${humanReadableScriptName} ***"
+
+            # Uninstall TelemetryDeck Client ID File
+            info "Uninstall TelemetryDeck Client ID File … "
+            tdClientIDFile="${organizationDirectory}/${organizationScriptName}TelemetryDeckClientID"
+            if [[ -f "${tdClientIDFile}" ]]; then
+                logComment "Removing '${tdClientIDFile}' … "
+                rm -f "${tdClientIDFile}" 2>&1
+                logComment "Removed '${tdClientIDFile}'"
+            else
+                logComment "TelemetryDeck Client ID file not found; nothing to remove"
+            fi
 
             # Uninstall LaunchDaemon
             info "Uninstall LaunchDaemon … "